+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Fixed
+- Fix dodola validate-dataset OOM on small workers without dask-distributed. (@brews)
 
 ## [0.17.0] - 2022-02-17
 ### Changed